Move declarations to their proper packages instead of just throwing them in a "declarations" package, more refactors
