+++ common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/commands/ApplicationCommandDeclarationWrappers.kt
@@ -1,0 +1,17 @@
+package net.perfectdreams.discordinteraktions.common.commands
+
+interface ApplicationCommandDeclarationWrapper {
+    fun declaration(): ApplicationCommandDeclaration
+}
+
+interface SlashCommandDeclarationWrapper : ApplicationCommandDeclarationWrapper {
+    override fun declaration(): SlashCommandDeclaration
+}
+
+interface UserCommandDeclarationWrapper : ApplicationCommandDeclarationWrapper{
+    override fun declaration(): UserCommandDeclaration
+}
+
+interface MessageCommandDeclarationWrapper : ApplicationCommandDeclarationWrapper{
+    override fun declaration(): MessageCommandDeclaration
+}
