+++ common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/commands/options/CommandOptionType.kt
@@ -1,4 +1,4 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.slash.options
+package net.perfectdreams.discordinteraktions.common.commands.options
 
 interface CommandOptionType {
     val isNullable: Boolean
