--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/commands/slash/SlashCommandExecutor.kt
@@ -1,13 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.commands.slash
-
-import net.perfectdreams.discordinteraktions.common.commands.interaction.InteractionCommandExecutor
-import net.perfectdreams.discordinteraktions.common.context.commands.ApplicationCommandContext
-import net.perfectdreams.discordinteraktions.common.context.commands.slash.SlashCommandArguments
-
-/**
- * This is the class that should be inherited if you
- * want to create an Slash Command.
- */
-abstract class SlashCommandExecutor : InteractionCommandExecutor() {
-    abstract suspend fun execute(context: ApplicationCommandContext, args: SlashCommandArguments)
-}
