--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/ComponentExecutor.kt
@@ -1,12 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components
-
-interface ComponentExecutor {
-    /**
-     * Used by the [net.perfectdreams.discordinteraktions.declarations.slash.SlashCommandExecutorDeclaration] to match declarations to executors.
-     *
-     * By default the class of the executor is used, but this may cause issues when using anonymous classes!
-     *
-     * To avoid this issue, you can replace the signature with another unique identifier
-     */
-    open fun signature(): Any = this::class
-}
