--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/buttons/ButtonClickExecutor.kt
@@ -1,5 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.buttons
-
-import net.perfectdreams.discordinteraktions.common.components.ComponentExecutor
-
-interface ButtonClickExecutor : ComponentExecutor
