--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/buttons/ButtonClickExecutorDeclaration.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.buttons
-
-import net.perfectdreams.discordinteraktions.common.components.ComponentDeclaration
-
-open class ButtonClickExecutorDeclaration(
-    parent: Any,
-    id: String
-) : ComponentDeclaration(parent, id)
