--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/buttons/ButtonClickWithDataExecutor.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.buttons
-
-import net.perfectdreams.discordinteraktions.api.entities.User
-import net.perfectdreams.discordinteraktions.common.context.components.ComponentContext
-
-interface ButtonClickWithDataExecutor : ButtonClickExecutor {
-    suspend fun onClick(user: User, context: ComponentContext, data: String)
-}
