--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/selects/SelectMenuExecutor.kt
@@ -1,5 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.selects
-
-import net.perfectdreams.discordinteraktions.common.components.ComponentExecutor
-
-interface SelectMenuExecutor : ComponentExecutor
