--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/selects/SelectMenuExecutorDeclaration.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.selects
-
-import net.perfectdreams.discordinteraktions.common.components.ComponentDeclaration
-
-open class SelectMenuExecutorDeclaration(
-    parent: Any,
-    id: String
-) : ComponentDeclaration(parent, id)
