--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/components/selects/SelectMenuWithDataExecutor.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.common.components.selects
-
-import net.perfectdreams.discordinteraktions.api.entities.User
-import net.perfectdreams.discordinteraktions.common.context.components.ComponentContext
-
-interface SelectMenuWithDataExecutor : SelectMenuExecutor {
-    suspend fun onSelect(user: User, context: ComponentContext, data: String, values: List<String>)
-}
