+++ common/src/main/kotlin/net/perfectdreams/discordinteraktions/common/context/commands/slash/SlashCommandArguments.kt
@@ -1,7 +1,7 @@
 package net.perfectdreams.discordinteraktions.common.context.commands.slash
 
-import net.perfectdreams.discordinteraktions.declarations.commands.slash.options.CommandOption
-import net.perfectdreams.discordinteraktions.declarations.commands.slash.options.CommandOptionType
+import net.perfectdreams.discordinteraktions.common.commands.options.CommandOption
+import net.perfectdreams.discordinteraktions.common.commands.options.CommandOptionType
 
 class SlashCommandArguments(val types: Map<CommandOption<*>, Any?>) {
     operator fun <T> get(argument: CommandOption<T>): T {
