--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/InteractionCommandDeclaration.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands
-
-/**
- * Base class of every interaction declaration, because all interactions share a [name]
- */
-sealed class InteractionCommandDeclaration(
-    val name: String
-)
