--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/MessageCommandDeclaration.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands
-
-import net.perfectdreams.discordinteraktions.declarations.commands.message.MessageCommandExecutorDeclaration
-
-class MessageCommandDeclaration(
-    name: String,
-    val executor: MessageCommandExecutorDeclaration // User/Message commands always requires an executor, that's why it is not nullable!
-) : InteractionCommandDeclaration(name)
