--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/SlashCommandDeclaration.kt
@@ -1,11 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands
-
-import net.perfectdreams.discordinteraktions.declarations.commands.slash.SlashCommandExecutorDeclaration
-
-class SlashCommandDeclaration(
-    name: String,
-    val description: String,
-    val executor: SlashCommandExecutorDeclaration? = null,
-    val subcommands: List<SlashCommandDeclaration>,
-    val subcommandGroups: List<SlashCommandGroupDeclaration>
-) : InteractionCommandDeclaration(name)
