--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/SlashCommandGroupDeclaration.kt
@@ -1,7 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands
-
-class SlashCommandGroupDeclaration(
-    name: String,
-    val description: String,
-    val subcommands: List<SlashCommandDeclaration>
-) : InteractionCommandDeclaration(name)
