--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/UserCommandDeclaration.kt
@@ -1,8 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands
-
-import net.perfectdreams.discordinteraktions.declarations.commands.user.UserCommandExecutorDeclaration
-
-class UserCommandDeclaration(
-    name: String,
-    val executor: UserCommandExecutorDeclaration // User/Message commands always requires an executor, that's why it is not nullable!
-) : InteractionCommandDeclaration(name)
