--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/application/ApplicationCommandExecutorDeclaration.kt
@@ -1,5 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.application
-
-// The "parent" is Any to avoid issues with anonymous classes
-// When using anonymous classes, you can use another type to match declarations
-open class ApplicationCommandExecutorDeclaration(val parent: Any)
