--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/wrappers/InteractionCommandDeclarationWrapper.kt
@@ -1,7 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.wrappers
-
-import net.perfectdreams.discordinteraktions.declarations.commands.InteractionCommandDeclaration
-
-interface InteractionCommandDeclarationWrapper {
-    fun declaration(): InteractionCommandDeclaration
-}
