--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/wrappers/MessageCommandDeclarationWrapper.kt
@@ -1,7 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.wrappers
-
-import net.perfectdreams.discordinteraktions.declarations.commands.MessageCommandDeclaration
-
-interface MessageCommandDeclarationWrapper : InteractionCommandDeclarationWrapper {
-    override fun declaration(): MessageCommandDeclaration
-}
