--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/wrappers/SlashCommandDeclarationWrapper.kt
@@ -1,7 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.wrappers
-
-import net.perfectdreams.discordinteraktions.declarations.commands.SlashCommandDeclaration
-
-interface SlashCommandDeclarationWrapper : InteractionCommandDeclarationWrapper {
-    override fun declaration(): SlashCommandDeclaration
-}
