--- common/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/wrappers/UserCommandDeclarationWrapper.kt
@@ -1,7 +1,0 @@
-package net.perfectdreams.discordinteraktions.declarations.commands.wrappers
-
-import net.perfectdreams.discordinteraktions.declarations.commands.UserCommandDeclaration
-
-interface UserCommandDeclarationWrapper : InteractionCommandDeclarationWrapper {
-    override fun declaration(): UserCommandDeclaration
-}
