+++ platforms/webserver-ktor-kord/src/samples/kotlin/SelectExecutor.kt
@@ -1,6 +1,6 @@
 import net.perfectdreams.discordinteraktions.api.entities.User
-import net.perfectdreams.discordinteraktions.common.components.selects.SelectMenuExecutorDeclaration
-import net.perfectdreams.discordinteraktions.common.components.selects.SelectMenuWithDataExecutor
+import net.perfectdreams.discordinteraktions.common.components.SelectMenuExecutorDeclaration
+import net.perfectdreams.discordinteraktions.common.components.SelectMenuWithDataExecutor
 import net.perfectdreams.discordinteraktions.common.context.components.ComponentContext
 
 class SelectExecutor : SelectMenuWithDataExecutor {
