+++ platforms/webserver-ktor-kord/src/samples/kotlin/TestClickExecutor.kt
@@ -2,11 +2,11 @@ import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import net.perfectdreams.discordinteraktions.api.entities.User
 import net.perfectdreams.discordinteraktions.common.builder.message.allowedMentions
-import net.perfectdreams.discordinteraktions.common.components.buttons.ButtonClickExecutorDeclaration
-import net.perfectdreams.discordinteraktions.common.components.buttons.ButtonClickWithNoDataExecutor
+import net.perfectdreams.discordinteraktions.common.components.ButtonClickExecutor
+import net.perfectdreams.discordinteraktions.common.components.ButtonClickExecutorDeclaration
 import net.perfectdreams.discordinteraktions.common.context.components.ComponentContext
 
-class TestClickExecutor : ButtonClickWithNoDataExecutor {
+class TestClickExecutor : ButtonClickExecutor {
     companion object : ButtonClickExecutorDeclaration(TestClickExecutor::class, "test_click")
 
     var mutex = Mutex()
